Refactor audio safety module: type fixes, constant relocation, and minor improvements
